Force page number int

+++ app/Api/Search/Services/SearchService.php
@@ -72,7 +72,7 @@ class SearchService
             'count' => $results->count(),
             'per_page' => $query->getParam('size'),
             'current_page' => $page,
-            'total_pages' =>  $totalPages <= 0 ? 1 : $totalPages
+            'total_pages' => (int) ($totalPages <= 0 ? 1 : $totalPages)
         ];
         return $pagination;
     }
